✨ [Feat] custom Hooks
