+++ week4/src/components/DayWeatherCard.jsx
@@ -1,43 +1,12 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import { useParams } from "react-router-dom";
-import axios from "axios";
-
 import { St } from "./CardStyle";
-
-import { WEATER_TYPE } from "../constants/WeatherType";
+import { useAxios } from "../lib/useAxios";
 
 const DayWeatherCard = () => {
   const { area } = useParams();
-  const [weather, setWeather] = useState({});
-  const [isLoading, setIsLoading] = useState(false);
 
-  const weatherInfo = () => {
-    try {
-      setIsLoading(true);
-      axios
-        .get(
-          `https://api.openweathermap.org/data/2.5/weather?q=${area}&appid=${
-            import.meta.env.VITE_APP_WEATHER
-          }&units=metric`
-        )
-        .then((response) => {
-          setWeather(response.data);
-          setIsLoading(false);
-        });
-    } catch (err) {
-      console.log(err);
-    }
-  };
-
-  const weatherDescription =
-    weather.weather?.length > 0 ? weather.weather[0].description : "";
-  const imgSrc = WEATER_TYPE.find(
-    (data) => data.description === weatherDescription
-  );
-
-  useEffect(() => {
-    weatherInfo();
-  }, [area]);
+  const { weather, isLoading, imgSrc } = useAxios(area);
 
   return (
     <>
